split up delta_scan.cpp
