+++ CMakeLists.txt
@@ -16,7 +16,9 @@ set(EXTENSION_SOURCES
     src/delta_functions.cpp
     src/delta_macros.cpp
     src/delta_utils.cpp
-    src/functions/delta_scan.cpp
+    src/functions/delta_scan/delta_scan.cpp
+    src/functions/delta_scan/delta_multi_file_list.cpp
+    src/functions/delta_scan/delta_multi_file_reader.cpp
     src/functions/expression_functions.cpp
     src/storage/delta_catalog.cpp
     src/storage/delta_schema_entry.cpp
