+++ src/delta_functions.cpp
@@ -2,7 +2,6 @@
 
 #include "duckdb.hpp"
 #include "duckdb/main/extension_util.hpp"
-#include "duckdb/parser/parsed_data/create_scalar_function_info.hpp"
 
 namespace duckdb {
