+++ src/delta_macros.cpp
@@ -1,12 +1,11 @@
 #include "delta_macros.hpp"
 
 #include "duckdb.hpp"
-#include "duckdb/main/extension_util.hpp"
-#include "duckdb/parser/parsed_data/create_scalar_function_info.hpp"
-#include "duckdb/parser/parser.hpp"
-#include "duckdb/function/table_macro_function.hpp"
-#include "duckdb/parser/parsed_data/create_macro_info.hpp"
 #include "duckdb/catalog/default/default_functions.hpp"
+#include "duckdb/main/extension_util.hpp"
+#include "duckdb/function/table_macro_function.hpp"
+#include "duckdb/parser/parser.hpp"
+#include "duckdb/parser/parsed_data/create_macro_info.hpp"
 
 namespace duckdb {
