+++ src/functions/expression_functions.cpp
@@ -1,6 +1,4 @@
-#include <duckdb/execution/expression_executor.hpp>
-#include <duckdb/function/built_in_functions.hpp>
-
+#include "duckdb/execution/expression_executor.hpp"
 #include "duckdb/function/scalar_function.hpp"
 #include "duckdb/planner/expression/bound_constant_expression.hpp"
