+++ src/storage/delta_catalog.cpp
@@ -6,8 +6,7 @@
 #include "duckdb/parser/parsed_data/create_schema_info.hpp"
 #include "duckdb/main/attached_database.hpp"
 
-#include "functions/delta_scan.hpp"
-#include "storage/delta_transaction_manager.hpp"
+#include "functions/delta_scan/delta_multi_file_list.hpp"
 
 namespace duckdb {
