+++ src/storage/delta_table_entry.cpp
@@ -1,11 +1,7 @@
+#include "functions/delta_scan/delta_scan.hpp"
 #include "storage/delta_catalog.hpp"
-#include "storage/delta_schema_entry.hpp"
 #include "storage/delta_table_entry.hpp"
 
-#include "delta_utils.hpp"
-#include "functions/delta_scan.hpp"
-
-#include "storage/delta_transaction.hpp"
 #include "duckdb/storage/statistics/base_statistics.hpp"
 #include "duckdb/storage/table_storage_info.hpp"
 #include "duckdb/main/extension_util.hpp"
@@ -13,9 +9,6 @@
 #include "duckdb/main/secret/secret_manager.hpp"
 #include "duckdb/catalog/catalog_entry/table_function_catalog_entry.hpp"
 #include "duckdb/parser/tableref/table_function_ref.hpp"
-#include "functions/delta_scan.hpp"
-
-#include <functional>
 
 namespace duckdb {
