+++ src/storage/delta_transaction.cpp
@@ -3,7 +3,7 @@
 #include "duckdb/parser/parsed_data/create_view_info.hpp"
 #include "duckdb/catalog/catalog_entry/index_catalog_entry.hpp"
 #include "duckdb/catalog/catalog_entry/view_catalog_entry.hpp"
-#include "functions/delta_scan.hpp"
+#include "functions/delta_scan/delta_scan.hpp"
 #include "storage/delta_table_entry.hpp"
 
 namespace duckdb {
